+++ esp-hal-procmacros/src/doc_replace.rs
@@ -6,6 +6,7 @@ use syn::{
     AttrStyle,
     Attribute,
     Expr,
+    ExprLit,
     Item,
     Lit,
     LitStr,
@@ -20,7 +21,16 @@ use syn::{
 };
 
 struct Replacements {
+    // Placeholder => [attribute contents]
+    //
+    // Replaces `# {tag}` placeholders with the attribute contents. Replaces the entire line.
     line_replacements: HashMap<String, Vec<TokenStream2>>,
+
+    // Placeholder => [(condition, string contents)], may be unconditional.
+    //
+    // Replaces `__tag__` placeholders with the string contents. Replaces only the placeholder
+    // inside the line, and applies the condition to the entire line if present.
+    inline_replacements: HashMap<String, Vec<(Option<TokenStream>, String)>>,
 }
 
 impl Replacements {
@@ -37,17 +47,28 @@ impl Replacements {
                         attrs.push(syn::parse_quote_spanned! { span => #![ #line ] });
                     }
                 }
+            } else if let Some((placeholder, replacements)) = self
+                .inline_replacements
+                .iter()
+                .find(|(k, _v)| trimmed.contains(k.as_str()))
+            {
+                for (cfg, replacement) in replacements.iter() {
+                    let line = line.replace(placeholder, replacement);
+                    let line = create_raw_string(&line);
+                    let attr_inner = if let Some(condition) = cfg {
+                        quote! { cfg_attr(#condition, doc = #line) }
+                    } else {
+                        quote! { doc = #line }
+                    };
+                    if outer {
+                        attrs.push(syn::parse_quote_spanned! { span => #[ #attr_inner ] });
+                    } else {
+                        attrs.push(syn::parse_quote_spanned! { span => #![ #attr_inner ] });
+                    }
+                }
             } else {
                 // Just append the line, in the expected format (`doc = r" Foobar"`)
-                let hash = if line.contains("#\"") {
-                    "##"
-                } else if line.contains('"') {
-                    "#"
-                } else {
-                    ""
-                };
-
-                let line = TokenStream2::from_str(&format!("r{hash}\"{line}\"{hash}")).unwrap();
+                let line = create_raw_string(line);
                 if outer {
                     attrs.push(syn::parse_quote_spanned! { span => #[doc = #line] });
                 } else {
@@ -61,35 +82,80 @@ impl Replacements {
     }
 }
 
+fn create_raw_string(line: &str) -> TokenStream2 {
+    let hash = if line.contains("#\"") {
+        "##"
+    } else if line.contains('"') {
+        "#"
+    } else {
+        ""
+    };
+
+    TokenStream2::from_str(&format!("r{hash}\"{line}\"{hash}")).unwrap()
+}
+
 impl Parse for Replacements {
     fn parse(input: syn::parse::ParseStream) -> syn::Result<Self> {
-        let mut replacements = HashMap::new();
+        let mut line_replacements = HashMap::new();
+        let mut inline_replacements = HashMap::new();
+
+        let mut add_line_replacement = |placeholder: &str, replacement: Vec<TokenStream2>| {
+            line_replacements.insert(format!("# {{{placeholder}}}"), replacement);
+        };
+        let mut add_inline_replacement =
+            |placeholder: &str, replacement: Vec<(Option<TokenStream2>, String)>| {
+                // The placeholder must be a valid Rust identifier to keep rustfmt happy
+                inline_replacements.insert(format!("__{placeholder}__"), replacement);
+            };
+
         if !input.is_empty() {
             let args = Punctuated::<Replacement, Token![,]>::parse_terminated(input)?;
             for arg in args {
-                let replacement = match arg.replacement {
-                    ReplacementKind::Literal(expr) => vec![quote! {
-                        doc = #expr
-                    }],
+                match arg.replacement {
+                    ReplacementKind::Literal(expr) => {
+                        if let Expr::Lit(ExprLit {
+                            lit: Lit::Str(ref lit_str),
+                            ..
+                        }) = expr
+                        {
+                            add_inline_replacement(&arg.placeholder, vec![(None, lit_str.value())]);
+                        }
+
+                        add_line_replacement(
+                            &arg.placeholder,
+                            vec![quote! {
+                                doc = #expr
+                            }],
+                        );
+                    }
                     ReplacementKind::Choice(items) => {
-                        let mut branches = vec![];
+                        let mut conditions = vec![];
+                        let mut bodies = vec![];
+                        let mut lit_strs = vec![];
                         let mut cfgs = vec![];
 
                         for branch in items {
                             let body = branch.body;
+
+                            if let Expr::Lit(ExprLit {
+                                lit: Lit::Str(ref lit_str),
+                                ..
+                            }) = body
+                            {
+                                lit_strs.push(lit_str.value());
+                            }
+
                             match branch.condition {
                                 Some(Meta::List(cfg)) if cfg.path.is_ident("cfg") => {
                                     let condition = cfg.tokens;
 
                                     cfgs.push(condition.clone());
-                                    branches.push(quote! {
-                                        cfg_attr(#condition, doc = #body)
-                                    });
+                                    conditions.push(condition);
+                                    bodies.push(body);
                                 }
                                 None => {
-                                    branches.push(quote! {
-                                        cfg_attr(not(any( #(#cfgs),*) ), doc = #body)
-                                    });
+                                    conditions.push(quote! { not(any( #(#cfgs),*) ) });
+                                    bodies.push(body);
                                 }
                                 _ => {
                                     return Err(syn::Error::new(
@@ -97,19 +163,36 @@ impl Parse for Replacements {
                                         "Expected a cfg condition or catch-all condition using `_`",
                                     ));
                                 }
-                            };
+                            }
                         }
 
-                        branches
-                    }
-                };
+                        let branches = conditions
+                            .iter()
+                            .zip(bodies.iter())
+                            .map(|(condition, body)| {
+                                quote! {
+                                    cfg_attr(#condition, doc = #body)
+                                }
+                            })
+                            .collect::<Vec<_>>();
+                        add_line_replacement(&arg.placeholder, branches);
 
-                replacements.insert(arg.placeholder, replacement);
+                        if lit_strs.len() == bodies.len() {
+                            let branches = conditions
+                                .into_iter()
+                                .map(Some)
+                                .zip(lit_strs.into_iter())
+                                .collect::<Vec<_>>();
+                            add_inline_replacement(&arg.placeholder, branches);
+                        }
+                    }
+                }
             }
         }
 
         Ok(Self {
-            line_replacements: replacements,
+            line_replacements,
+            inline_replacements,
         })
     }
 }
@@ -126,7 +209,7 @@ impl Parse for Replacement {
         let replacement: ReplacementKind = input.parse()?;
 
         Ok(Self {
-            placeholder: format!("# {{{}}}", placeholder.value()),
+            placeholder: placeholder.value(),
             replacement,
         })
     }
@@ -391,6 +474,66 @@ mod tests {
                 /// ```rust, no_run
                 #[cfg_attr (esp32h2 , doc = "let freq = Rate::from_mhz(32);")]
                 #[cfg_attr (not (any (esp32h2)) , doc = "let freq = Rate::from_mhz(80);")]
+                #[doc = crate::before_snippet!()]
+                /// let peripherals = esp_hal::init(esp_hal::Config::default());
+                #[doc = crate::after_snippet!()]
+                /// ```
+                struct Foo {}
+            }
+            .to_string()
+        );
+    }
+
+    #[test]
+    fn test_custom_inline_replacements() {
+        let result = replace(
+            quote! {
+                "freq" => {
+                    cfg(esp32h2) => "32",
+                    _ => "80"
+                },
+                "other" => "Replacement"
+            }
+            .into(),
+            quote! {
+                /// # Configuration
+                /// ## Overview
+                /// This module contains the initial configuration for the system.
+                /// ## Configuration
+                /// In the [`esp_hal::init()`][crate::init] method, we can configure different
+                /// parameters for the system:
+                /// - CPU clock configuration.
+                /// - Watchdog configuration.
+                /// ## __other__ Examples
+                /// ### Default initialization
+                /// ```rust, no_run
+                /// let freq = Rate::from_mhz(__freq__);
+                /// # {before_snippet}
+                /// let peripherals = esp_hal::init(esp_hal::Config::default());
+                /// # {after_snippet}
+                /// ```
+                struct Foo {
+                }
+            }
+            .into(),
+        );
+
+        assert_eq!(
+            result.to_string(),
+            quote! {
+                /// # Configuration
+                /// ## Overview
+                /// This module contains the initial configuration for the system.
+                /// ## Configuration
+                /// In the [`esp_hal::init()`][crate::init] method, we can configure different
+                /// parameters for the system:
+                /// - CPU clock configuration.
+                /// - Watchdog configuration.
+                /// ## Replacement Examples
+                /// ### Default initialization
+                /// ```rust, no_run
+                #[cfg_attr (esp32h2 , doc = r" let freq = Rate::from_mhz(32);")]
+                #[cfg_attr (not (any (esp32h2)) , doc = r" let freq = Rate::from_mhz(80);")]
                 #[doc = crate::before_snippet!()]
                 /// let peripherals = esp_hal::init(esp_hal::Config::default());
                 #[doc = crate::after_snippet!()]