+++ esp-hal/src/lib.rs
@@ -5,14 +5,7 @@
 //! # Bare-metal (`no_std`) HAL for all Espressif ESP32 devices.
 //!
 //! This documentation is built for the
-#![cfg_attr(esp32, doc = "**ESP32**")]
-#![cfg_attr(esp32s2, doc = "**ESP32-S2**")]
-#![cfg_attr(esp32s3, doc = "**ESP32-S3**")]
-#![cfg_attr(esp32c2, doc = "**ESP32-C2**")]
-#![cfg_attr(esp32c3, doc = "**ESP32-C3**")]
-#![cfg_attr(esp32c5, doc = "**ESP32-C5**")]
-#![cfg_attr(esp32c6, doc = "**ESP32-C6**")]
-#![cfg_attr(esp32h2, doc = "**ESP32-H2**")]
+#![doc = concat!("**", chip_pretty!(), "**")]
 //! . Please ensure you are reading the correct [documentation] for your target
 //! device.
 //!