+++ esp-hal/src/rng/mod.rs
@@ -1,7 +1,7 @@
 #![cfg_attr(docsrs, procmacros::doc_replace(
     "analog_pin" => {
-        cfg(esp32) => "let analog_pin = peripherals.GPIO32;",
-        cfg(not(esp32)) => "let analog_pin = peripherals.GPIO3;"
+        cfg(esp32) => "GPIO32",
+        _ => "GPIO3"
     },
     "documentation" => concat!("[ESP-IDF documentation](https://docs.espressif.com/projects/esp-idf/en/latest/", chip!(), "/api-reference/system/random.html)")
 ))]
@@ -124,10 +124,8 @@ let rng = trng.downgrade();
 // Drop the true random number source. ADC is available now.
 core::mem::drop(trng_source);
 
-# {analog_pin}
-
 let mut adc1_config = AdcConfig::new();
-let mut adc1_pin = adc1_config.enable_pin(analog_pin, Attenuation::_11dB);
+let mut adc1_pin = adc1_config.enable_pin(peripherals.__analog_pin__, Attenuation::_11dB);
 let mut adc1 = Adc::<ADC1, Blocking>::new(peripherals.ADC1, adc1_config);
 let pin_value: u16 = nb::block!(adc1.read_oneshot(&mut adc1_pin))?;